Added template for when the user isn't authorized to perform the action

+++ itemcatalog.py
@@ -212,6 +212,8 @@ def showCategories():
 def newCategory():
     if 'username' not in login_session:
         return redirect('/login')
+    if login_session.get('is_admin') != True:
+        return render_template('notauthorized.html'), 403
     if request.method == 'POST':
         newCategory = Category(
             name=request.form['name'])
@@ -227,6 +229,8 @@ def newCategory():
 def editCategory(category_id):
     if 'username' not in login_session:
         return redirect('/login')
+    if login_session.get('is_admin') != True:
+        return render_template('notauthorized.html'), 403
     editedCategory = session.query(
         Category).filter_by(id=category_id).one()
     if request.method == 'POST':
@@ -244,6 +248,8 @@ def editCategory(category_id):
 def deleteCategory(category_id):
     if 'username' not in login_session:
         return redirect('/login')
+    if login_session.get('is_admin') != True:
+        return render_template('notauthorized.html'), 403
     categoryToDelete = session.query(
         Category).filter_by(id=category_id).one()
     if request.method == 'POST':
@@ -272,6 +278,7 @@ def showCategory(category_id):
 def newItem(category_id):
     if 'username' not in login_session:
         return redirect('/login')
+    # Any logged in user can create an item, so don't check user id
     category = session.query(
         Category).filter_by(id=category_id).one()
     if request.method == 'POST':
@@ -293,6 +300,9 @@ def editItem(item_id):
     if 'username' not in login_session:
         return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if login_session.get('is_admin') != True and (
+            login_session.get('user_id') is None or login_session.get('user_id') != item.user_id):
+        return render_template('notauthorized.html'), 403
     if request.method == 'POST':
         item.name = request.form['name']
         item.description = request.form['description']
@@ -309,6 +319,9 @@ def deleteItem(item_id):
     if 'username' not in login_session:
         return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if login_session.get('is_admin') != True and (
+            login_session.get('user_id') is None or login_session.get('user_id') != item.user_id):
+        return render_template('notauthorized.html'), 403
     if request.method == 'POST':
         session.delete(item)
         session.commit()
